Refactor foil overlay code.
