+++ public/css/draft.css
@@ -23,18 +23,6 @@
   margin: 0 -2px;
 }
 
-@media(min-width: 768px) {
-  .col-md-1-5 {
-    flex: 0 0 12.5%;
-    max-width: 12.5%;
-  }
-
-  .col-md-1-4285 {
-    flex: 0 0 14.285%;
-    max-width: 14.285%;
-  }
-}
-
 .off-screen {
   position: absolute;
   z-index: -10000;
