+++ src/components/FoilOverlay.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const FoilOverlay = (Tag) => (props) =>
+  <div>
+    <img
+      src="/content/foilOverlay.png"
+      className={'foilOverlay card-border' + (props.card && props.card.finish === 'Foil' ? '' : ' d-none')}
+      width="100%"
+    />
+    <Tag {...props} />
+  </div>;
+
+export default FoilOverlay;
