+++ src/components/VisualSpoiler.js
@@ -2,8 +2,8 @@ import React, { useEffect } from 'react';
 
 import { Row, Col } from 'reactstrap';
 
-import AutocardImage from './AutocardImage';
 import SortContext from './SortContext';
+import SpoilerImage from './SpoilerImage';
 
 const VisualSpoilerRaw = ({ cards, primary, secondary, tertiary, changeSort, ...props }) => {
   const groups = {};
@@ -25,10 +25,8 @@ const VisualSpoilerRaw = ({ cards, primary, secondary, tertiary, changeSort, ...
                 .filter((label3) => groups[label1][label2][label3])
                 .map((label3) =>
                   groups[label1][label2][label3].map(({ index, tags, finish, details }) => (
-                    <Col key={index} className="w-auto flex-grow-0">
-                      <div className="visualSpoilerCardContainer">
-                        <AutocardImage index={index} tags={tags} finish={finish} {...details} />
-                      </div>
+                    <Col key={index} xs={6} sm={4} className="col-md-1-5">
+                      <SpoilerImage index={index} tags={tags} finish={finish} {...details} />
                     </Col>
                   )),
                 ),
